requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
pydantic==2.10.4
python-dotenv==1.2.2
